Move camstack.viewers into viewertools and viewermain
